refactor(infra): move known abi fallback decoder logic
