+++ typescript/infra/src/tx/govern-transaction-reader.ts
@@ -10,9 +10,6 @@ import { BigNumber, ethers } from 'ethers';
 import {
   BaseFee__factory,
   CrossCollateralRouter__factory,
-  CrossCollateralRoutingFee__factory,
-  DomainRoutingHook__factory,
-  DomainRoutingIsm__factory,
   HypXERC20Lockbox__factory,
   MovableCollateralRouter__factory,
   Ownable__factory,
@@ -438,8 +435,6 @@ export class GovernTransactionReader {
         this.readWarpModuleTransaction(chain, tx),
       isFeeTransaction: (chain, tx) => this.isFeeTransaction(chain, tx),
       readFeeTransaction: (chain, tx) => this.readFeeTransaction(chain, tx),
-      tryReadByKnownContractInterface: (chain, tx) =>
-        this.tryReadByKnownContractInterface(chain, tx),
     };
   }
 
@@ -974,400 +969,6 @@ export class GovernTransactionReader {
       signature: decoded.signature,
       ...(feeDetails && { feeDetails }),
     };
-  }
-
-  /**
-   * Selector-based fallback for known Hyperlane contract types that may not be
-   * present in the warp route registry (e.g. newly-deployed warp routes,
-   * standalone routing hooks/ISMs, token bridge adapters, cross-collateral
-   * routing fee contracts). Returns undefined if no interface matches.
-   */
-  private tryReadByKnownContractInterface(
-    chain: ChainName,
-    tx: AnnotatedEV5Transaction,
-  ): GovernTransaction | undefined {
-    if (!tx.data || tx.data.length < 10 || !tx.to) return undefined;
-
-    const tryParse = (iface: ethers.utils.Interface) => {
-      try {
-        return iface.parseTransaction({ data: tx.data!, value: tx.value });
-      } catch {
-        return undefined;
-      }
-    };
-
-    const formatBase = (
-      contractType: string,
-      decoded: ethers.utils.TransactionDescription,
-      insight: string,
-    ) => ({
-      chain,
-      to: `${contractType} (${chain} ${tx.to})`,
-      insight,
-      signature: decoded.signature,
-      decoderMatch: {
-        confidence: 'selector-only',
-        insight: 'unverified ABI match',
-      },
-    });
-
-    // CrossCollateralRouter / MovableCollateralRouter (TokenRouter base) - newly
-    // deployed warp routes not yet in the registry.
-    const ccrIface = CrossCollateralRouter__factory.createInterface();
-    const movableIface = MovableCollateralRouter__factory.createInterface();
-    const ccrDecoded = tryParse(ccrIface) ?? tryParse(movableIface);
-    if (ccrDecoded) {
-      const insight = this.formatRouterCallInsight(
-        ccrIface,
-        movableIface,
-        ccrDecoded,
-      );
-      return formatBase('Warp Route (unregistered)', ccrDecoded, insight);
-    }
-
-    // TokenBridgeOft adapter (LayerZero OFT bridge for warp routes).
-    const oftIface = TokenBridgeOft__factory.createInterface();
-    const oftDecoded = tryParse(oftIface);
-    if (oftDecoded) {
-      const insight = this.formatTokenBridgeOftInsight(oftIface, oftDecoded);
-      return formatBase('TokenBridgeOft', oftDecoded, insight);
-    }
-
-    // TokenBridgeDepositAddress adapter (deposit-address-based bridges).
-    const depositAddrIface =
-      TokenBridgeDepositAddress__factory.createInterface();
-    const depositAddrDecoded = tryParse(depositAddrIface);
-    if (depositAddrDecoded) {
-      const insight = this.formatTokenBridgeDepositAddressInsight(
-        depositAddrIface,
-        depositAddrDecoded,
-      );
-      return formatBase(
-        'TokenBridgeDepositAddress',
-        depositAddrDecoded,
-        insight,
-      );
-    }
-
-    // CrossCollateralRoutingFee - per-router fee contract routing.
-    const ccrFeeIface = CrossCollateralRoutingFee__factory.createInterface();
-    const ccrFeeDecoded = tryParse(ccrFeeIface);
-    if (ccrFeeDecoded) {
-      const insight = this.formatCrossCollateralRoutingFeeInsight(
-        ccrFeeIface,
-        ccrFeeDecoded,
-      );
-      return formatBase('CrossCollateralRoutingFee', ccrFeeDecoded, insight);
-    }
-
-    // DomainRoutingHook.setHook / setHooks
-    const routingHookIface = DomainRoutingHook__factory.createInterface();
-    const routingHookDecoded = tryParse(routingHookIface);
-    if (routingHookDecoded) {
-      const insight = this.formatDomainRoutingHookInsight(
-        routingHookIface,
-        routingHookDecoded,
-      );
-      return formatBase('DomainRoutingHook', routingHookDecoded, insight);
-    }
-
-    // DomainRoutingIsm.set / remove
-    const routingIsmIface = DomainRoutingIsm__factory.createInterface();
-    const routingIsmDecoded = tryParse(routingIsmIface);
-    if (routingIsmDecoded) {
-      const insight = this.formatDomainRoutingIsmInsight(
-        routingIsmIface,
-        routingIsmDecoded,
-      );
-      return formatBase('DomainRoutingIsm', routingIsmDecoded, insight);
-    }
-
-    return undefined;
-  }
-
-  private formatRouterCallInsight(
-    ccrIface: ethers.utils.Interface,
-    movableIface: ethers.utils.Interface,
-    decoded: ethers.utils.TransactionDescription,
-  ): string {
-    const args = decoded.args;
-
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'addBridge(uint32,address)',
-      )
-    ) {
-      return `Set bridge for origin domain ${this.formatDomain(args[0])} to ${args[1]}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'removeBridge(uint32,address)',
-      )
-    ) {
-      return `Remove bridge ${args[1]} from domain ${this.formatDomain(args[0])}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'enrollRemoteRouter(uint32,bytes32)',
-      )
-    ) {
-      return `Enroll remote router for domain ${this.formatDomain(args[0])} to ${args[1]}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'enrollRemoteRouters(uint32[],bytes32[])',
-      )
-    ) {
-      const [domains, routers] = args;
-      const lines = domains.map(
-        (d: number, i: number) =>
-          `domain ${this.formatDomain(d)} to ${routers[i]}`,
-      );
-      return `Enroll remote routers for ${lines.join(', ')}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'unenrollRemoteRouter(uint32)',
-      )
-    ) {
-      return `Unenroll remote router for domain ${this.formatDomain(args[0])}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'unenrollRemoteRouters(uint32[])',
-      )
-    ) {
-      const lines = args[0].map(
-        (d: number) => `domain ${this.formatDomain(d)}`,
-      );
-      return `Unenroll remote routers for ${lines.join(', ')}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'setDestinationGas((uint32,uint256)[])',
-      )
-    ) {
-      const lines = args[0].map(
-        (c: { domain: number; gas: BigNumber }) =>
-          `domain ${this.formatDomain(c.domain)} to ${c.gas.toString()}`,
-      );
-      return `Set destination gas for ${lines.join(', ')}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'setDestinationGas(uint32,uint256)',
-      )
-    ) {
-      return `Set destination gas for domain ${this.formatDomain(args[0])} to ${args[1].toString()}`;
-    }
-    if (matchesFunctionSignature(decoded, movableIface, 'setHook(address)')) {
-      return `Set hook to ${args[0]}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'setInterchainSecurityModule(address)',
-      )
-    ) {
-      return `Set ISM to ${args[0]}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'setFeeRecipient(address)',
-      )
-    ) {
-      return `Set fee recipient to ${args[0]}`;
-    }
-    if (
-      matchesFunctionSignature(decoded, movableIface, 'addRebalancer(address)')
-    ) {
-      return `Add rebalancer ${args[0]}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'removeRebalancer(address)',
-      )
-    ) {
-      return `Remove rebalancer ${args[0]}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'setRecipient(uint32,bytes32)',
-      )
-    ) {
-      return `Set rebalance recipient for domain ${this.formatDomain(args[0])} to ${args[1]}`;
-    }
-    if (
-      matchesFunctionSignature(decoded, movableIface, 'removeRecipient(uint32)')
-    ) {
-      return `Remove rebalance recipient for domain ${this.formatDomain(args[0])}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        movableIface,
-        'approveTokenForBridge(address,address)',
-      )
-    ) {
-      return `Approve token ${args[0]} for bridge ${args[1]}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        ccrIface,
-        'enrollCrossCollateralRouters(uint32[],bytes32[])',
-      )
-    ) {
-      const [domains, routers] = args;
-      const lines = domains.map(
-        (d: number, i: number) =>
-          `domain ${this.formatDomain(d)} to ${routers[i]}`,
-      );
-      return `Enroll cross-collateral routers for ${lines.join(', ')}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        ccrIface,
-        'unenrollCrossCollateralRouters(uint32[],bytes32[])',
-      )
-    ) {
-      const [domains, routers] = args;
-      const lines = domains.map(
-        (d: number, i: number) =>
-          `domain ${this.formatDomain(d)} from ${routers[i]}`,
-      );
-      return `Unenroll cross-collateral routers for ${lines.join(', ')}`;
-    }
-    return `Call ${decoded.signature}`;
-  }
-
-  private formatTokenBridgeOftInsight(
-    iface: ethers.utils.Interface,
-    decoded: ethers.utils.TransactionDescription,
-  ): string {
-    const args = decoded.args;
-    if (matchesFunctionSignature(decoded, iface, 'addDomain(uint32,uint32)')) {
-      return `Map Hyperlane domain ${this.formatDomain(args[0])} to LayerZero EID ${args[1]}`;
-    }
-    if (matchesFunctionSignature(decoded, iface, 'removeDomain(uint32)')) {
-      return `Remove Hyperlane domain ${this.formatDomain(args[0])} mapping`;
-    }
-    if (matchesFunctionSignature(decoded, iface, 'setExtraOptions(bytes)')) {
-      return `Set extra LayerZero options`;
-    }
-    return `Call ${decoded.signature}`;
-  }
-
-  private formatTokenBridgeDepositAddressInsight(
-    iface: ethers.utils.Interface,
-    decoded: ethers.utils.TransactionDescription,
-  ): string {
-    const args = decoded.args;
-    if (
-      matchesFunctionSignature(
-        decoded,
-        iface,
-        'addDestinationConfig(uint32,address,bytes32,uint256)',
-      )
-    ) {
-      return `Add destination config: domain ${this.formatDomain(args[0])}, depositAddress ${args[1]}, recipient ${args[2]}, feeBps ${args[3].toString()}`;
-    }
-    if (
-      matchesFunctionSignature(
-        decoded,
-        iface,
-        'removeDestinationConfig(uint32,bytes32)',
-      )
-    ) {
-      return `Remove destination config: domain ${this.formatDomain(args[0])}, recipient ${args[1]}`;
-    }
-    return `Call ${decoded.signature}`;
-  }
-
-  private formatCrossCollateralRoutingFeeInsight(
-    iface: ethers.utils.Interface,
-    decoded: ethers.utils.TransactionDescription,
-  ): string {
-    const args = decoded.args;
-    if (
-      matchesFunctionSignature(
-        decoded,
-        iface,
-        'setCrossCollateralRouterFeeContracts(uint32[],bytes32[],address[])',
-      )
-    ) {
-      const [destinations, targetRouters, feeContracts] = args;
-      const lines = destinations.map(
-        (d: number, i: number) =>
-          `domain ${this.formatDomain(d)} router ${targetRouters[i]} → fee ${feeContracts[i]}`,
-      );
-      return `Set per-router fee contracts: ${lines.join(', ')}`;
-    }
-    if (matchesFunctionSignature(decoded, iface, 'claim(address,address)')) {
-      return `Claim ${args[1]} balance to ${args[0]}`;
-    }
-    return `Call ${decoded.signature}`;
-  }
-
-  private formatDomainRoutingHookInsight(
-    iface: ethers.utils.Interface,
-    decoded: ethers.utils.TransactionDescription,
-  ): string {
-    const args = decoded.args;
-    if (matchesFunctionSignature(decoded, iface, 'setHook(uint32,address)')) {
-      return `Set hook for destination ${this.formatDomain(args[0])} to ${args[1]}`;
-    }
-    if (matchesFunctionSignature(decoded, iface, 'setHook(address)')) {
-      return `Set mailbox client hook to ${args[0]}`;
-    }
-    if (
-      matchesFunctionSignature(decoded, iface, 'setHooks((uint32,address)[])')
-    ) {
-      const lines = args[0].map(
-        (cfg: { destination: number; hook: string }) =>
-          `destination ${this.formatDomain(cfg.destination)} → ${cfg.hook}`,
-      );
-      return `Set hooks: ${lines.join(', ')}`;
-    }
-    return `Call ${decoded.signature}`;
-  }
-
-  private formatDomainRoutingIsmInsight(
-    iface: ethers.utils.Interface,
-    decoded: ethers.utils.TransactionDescription,
-  ): string {
-    const args = decoded.args;
-    if (matchesFunctionSignature(decoded, iface, 'set(uint32,address)')) {
-      return `Set ISM for origin ${this.formatDomain(args[0])} to ${args[1]}`;
-    }
-    if (matchesFunctionSignature(decoded, iface, 'remove(uint32)')) {
-      return `Remove ISM for origin ${this.formatDomain(args[0])}`;
-    }
-    return `Call ${decoded.signature}`;
   }
 
   /**
